Fix bug in ForceTerminate on Linux/ppc
On bruckner, a ForceTerminate will cause pre-Exit and pre-LWP destroy
events to be created in response to the terminate. Due to some
OS-level race condition, the OS will deliver a post-Exit for the
process but it will still exist in a stopped state. Continues do
not cause the process to exit.

The workaround for this bug is unset ptrace event options for
the process before terminate so we will not get the pre-Exit and
pre-LWP destroy events. This appears to avoid the case where the
process will be stopped following a ForceTerminate.

+++ proccontrol/src/process.C
@@ -955,6 +955,10 @@ bool int_process::terminate(bool &needs_sync)
    return !had_error;
 }
 
+bool int_process::preTerminate() {
+    return true;
+}
+
 int_process::int_process(Dyninst::PID p, std::string e,
                          std::vector<std::string> a,
                          std::vector<std::string> envp,
@@ -4483,6 +4487,13 @@ bool Process::terminate()
    }
 
    pthrd_printf("User terminating process %d\n", llproc_->getPid());
+
+   if( !llproc_->preTerminate() ) {
+       perr_printf("pre-terminate hook failed\n");
+       setLastError(err_internal, "Pre-terminate hook failed\n");
+       return false;
+   }
+
    bool needsSync = false;
    bool result = llproc_->terminate(needsSync);
    if (!result) {
